Moved analysis ajax requests into a separate controller. Updated urls. Added keys to Analysis react components within a router
