+++ src/main/webapp/pages/projects/analyses/_base.html
@@ -4,6 +4,7 @@
 <head>
     <title th:text="${project.label} + ' - ' + #{project.nav.analysis}"></title>
     <link rel="stylesheet" th:href="@{/dist/css/analysis.bundle.css}">
+    <link rel="stylesheet" th:href="@{/dist/css/analysis-outputs-table.bundle.css}">
 </head>
 <body>
 <main layout:fragment="main">
