+++ src/main/webapp/resources/js/apis/analysis/analysis.js
@@ -4,7 +4,7 @@
 import axios from "axios";
 
 // Ajax URL for Analysis
-const URL = `${window.TL.BASE_URL}analysis/ajax/`;
+const URL = `${window.TL.BASE_URL}ajax/analysis`;
 const BASE_URL = `${window.TL.BASE_URL}ajax/analyses`;
 
 /*
@@ -14,8 +14,8 @@ const BASE_URL = `${window.TL.BASE_URL}ajax/analyses`;
  *                      `error` contains error information if an error occurred.
  */
 export async function getVariablesForDetails(submissionId) {
-  const res = await axios.get(`${URL}details/${submissionId}`);
-  return res.data.analysisDetails;
+  const { data } = await axios.get(`${URL}/details/${submissionId}`);
+  return data;
 }
 
 /*
@@ -26,8 +26,8 @@ export async function getVariablesForDetails(submissionId) {
  */
 export async function getAnalysisInputFiles(submissionId) {
   try {
-    const res = await axios.get(`${URL}inputs/${submissionId}`);
-    return res.data.analysisInputFiles;
+    const { data } = await axios.get(`${URL}/inputs/${submissionId}`);
+    return data;
   } catch (error) {
     return { samples: [], referenceFile: null };
   }
@@ -42,13 +42,13 @@ export async function getAnalysisInputFiles(submissionId) {
  */
 export async function updateAnalysisEmailPipelineResult(params) {
   try {
-    const res = await axios.patch(`${URL}update-email-pipeline-result`, {
+    const { data } = await axios.patch(`${URL}/update-email-pipeline-result`, {
       analysisSubmissionId: params.submissionId,
       emailPipelineResult: params.emailPipelineResult
     });
-    return res.data.responseDetails.message;
+    return data.message;
   } catch (error) {
-    return { text: error.response.data.responseDetails.message, type: "error" };
+    return { text: error.response.data.message, type: "error" };
   }
 }
 
@@ -61,14 +61,14 @@ export async function updateAnalysisEmailPipelineResult(params) {
  */
 export async function updateAnalysis(params) {
   try {
-    const res = await axios.patch(`${URL}update-analysis/`, {
+    const { data } = await axios.patch(`${URL}/update-analysis/`, {
       analysisSubmissionId: params.submissionId,
       analysisName: params.analysisName,
       priority: params.priority
     });
-    return res.data.responseDetails.message;
+    return data.message;
   } catch (error) {
-    return { text: error.response.data.responseDetails.message, type: "error" };
+    return { text: error.response.data.message, type: "error" };
   }
 }
 
@@ -78,8 +78,8 @@ export async function updateAnalysis(params) {
  * @return {Promise<*>} `data` contains the OK response; `error` contains error information if an error occurred.
  */
 export async function deleteAnalysis(submissionId) {
-  const res = await axios.delete(`${URL}delete/${submissionId}`);
-  return res.data;
+  const { data } = await axios.delete(`${URL}/delete/${submissionId}`);
+  return data;
 }
 
 /*
@@ -88,8 +88,8 @@ export async function deleteAnalysis(submissionId) {
  * @return {Promise<*>} `data` contains the OK response; `error` contains error information if an error occurred.
  */
 export async function getSharedProjects(submissionId) {
-  const res = await axios.get(`${URL}${submissionId}/share`);
-  return res.data;
+  const { data } = await axios.get(`${URL}/${submissionId}/share`);
+  return data;
 }
 
 /*
@@ -100,11 +100,11 @@ export async function getSharedProjects(submissionId) {
  * @return {Promise<*>} `data` contains the OK response; `error` contains error information if an error occurred.
  */
 export async function updateSharedProject(params) {
-  const res = await axios.post(`${URL}${params.submissionId}/share`, {
+  const { data } = await axios.post(`${URL}/${params.submissionId}/share`, {
     projectId: params.projectId,
     shareStatus: params.shareStatus
   });
-  return res.data;
+  return data;
 }
 
 /*
@@ -114,8 +114,8 @@ export async function updateSharedProject(params) {
  */
 export async function saveToRelatedSamples(submissionId) {
   try {
-    const res = await axios.post(`${URL}${submissionId}/save-results`);
-    return res.data.message;
+    const { data } = await axios.post(`${URL}/${submissionId}/save-results`);
+    return data.message;
   } catch (error) {
     return { text: error.response.data.message, type: "error" };
   }
@@ -128,10 +128,8 @@ export async function saveToRelatedSamples(submissionId) {
  */
 export async function getJobErrors(submissionId) {
   try {
-    const res = await axios.get(
-      `${window.TL.BASE_URL}analysis/ajax/${submissionId}/job-errors`
-    );
-    return res.data;
+    const { data } = await axios.get(`${URL}/${submissionId}/job-errors`);
+    return data;
   } catch (error) {
     return { error: error };
   }
@@ -144,8 +142,8 @@ export async function getJobErrors(submissionId) {
  */
 export async function getSistrResults(submissionId) {
   try {
-    const res = await axios.get(`${URL}sistr/${submissionId}`);
-    return res.data;
+    const { data } = await axios.get(`${URL}/sistr/${submissionId}`);
+    return data;
   } catch (error) {
     return { error: error };
   }
@@ -172,7 +170,7 @@ export async function getPrincipalUserSingleSampleAnalysisOutputs() {
 export async function getProjectSharedSingleSampleAnalysisOutputs(projectId) {
   try {
     const { data } = await axios.get(
-      `${URL}project/${projectId}/shared-analysis-outputs`
+      `${URL}/project/${projectId}/shared-analysis-outputs`
     );
     return { data };
   } catch (error) {
@@ -190,7 +188,7 @@ export async function getProjectAutomatedSingleSampleAnalysisOutputs(
 ) {
   try {
     const { data } = await axios.get(
-      `${URL}project/${projectId}/automated-analysis-outputs`
+      `${URL}/project/${projectId}/automated-analysis-outputs`
     );
     return { data };
   } catch (error) {
@@ -207,7 +205,7 @@ export async function prepareAnalysisOutputsDownload(outputs) {
   try {
     const { data } = await axios({
       method: "post",
-      url: `${BASE_URL}download/prepare`,
+      url: `${URL}/download/prepare`,
       data: outputs
     });
     return { data };
