+++ src/main/webapp/resources/js/pages/analysis/json-preview.js
@@ -195,7 +195,7 @@ function jsToHtml(x, acc = "") {
 /**
  * Render a preview of a JSON AnalysisOutputFile
  * @param {jQuery|HTMLElement} $container Container element to render preview in
- * @param {string} baseUrl Base AJAX URL (e.g. /analysis/ajax/)
+ * @param {string} baseUrl Base AJAX URL (e.g. /ajax/analysis/)
  * @param {Object} aof AnalysisOutputFile info
  * @param {number} height Preview container height
  * @param {number} chunk_size Number of bytes to read from AnalysisOutputFile at a time
